Another bug fix

+++ coord.py
@@ -2,7 +2,7 @@
 
 def coord_for(n, a=0, b=1):
     coords = []
-    dist = (b-a)/n
+    dist = float((b-a)/n)
     #print(dist)
     for i in range(n+1):
         coords.append(a+i*dist)
@@ -11,7 +11,7 @@ def coord_for(n, a=0, b=1):
 def coord_while(n, a=0, b=1):
     i = 0
     coords = []
-    dist = (b-a)/n
+    dist = float((b-a)/n)
     #print(dist)
     while((a+i)<=b):
         coords.append(a+i)
@@ -19,7 +19,7 @@ def coord_while(n, a=0, b=1):
     return coords
 
 def coord_comp(n, a=0, b=1):
-    dist = (b-a)/n
+    dist = float((b-a)/n)
     #print(dist)
     coords = [a+i*dist for i in range(n+1)]
     return coords
